// src/app/models/register.model.ts
export class Register {
  
    id: number = 0;
    first_name: string = '';
    last_name: string = '';
    email: string = '';
    country_code: string = '';
    contact_number: string = '' ;
    user_status: string = '';
    password: string = ' ';
    type: string = ''; // Add the missing "type" field if required by backend

    
  }
  